Add Frostsnap app aka Frostsnapp (#80)
* [coord] s/unique_id/id/g

* Make OpenError Display rather than std::error::Error

Forcing it to be std::error::Error is not idiomatic apparently. I only
need it to be `Display` really.

* Prepare for frostsnap app

- Rearranged serial interfaces to that they were as simple as possible
while being flexible enough for what we needed on android side.
- Made poll_ports return a kind of changeset which makes notifying UI of
changes easier and actually makes the cli workflow more robust as well.
- We strictly set the interface to UART mode first so we don't have the
JTAG interface appear to the OS until it's its turn.

* Add "frostsnapp" flutter app

works on linux and android. Might work on windows and MacOS took.
It only allows you to register devices (give them names).

* Set frostypede ports to start in uart

* Upgrade flutter_rust_bridge version to 1.81.0

* Update instructions with Nick's feedback

* Slow down magic bytes on coordinator

To the same speed that the devices write downstream.
Also do some refactoring of the app.

* [ci] Futher simplify CI

We need libudev even to do clippy

* [ci] Use justfile

And fix all the clippy problems this found.

* [ci] Try building app in ci

* Improve README with all the things I learned

* NDK install steps in README

* Remove license and add to readme

* [ci] run apt-get update

* fix flutter_rust_bridge dart dep

---------

Co-authored-by: Adam Mashrique <amashrique@gmail.com>
Co-authored-by: Nick Farrow <nick@nickfarrow.com>