[app] Make it a Stream<String> rather than LogEntry
To centralise the log formatting on the rust side rather than doing it
in multiple places. We can now use tracing's tools to control
formatting.

Downside is you have to regex the log in order to color it. Upside
you've got the time.

We also include the logs when we have a STARTUP ERROR.