[comms] You can only send a message when you have the conch 🐚
This system makes it unlikely that a device can be busy doing something
else when a downstream device sends a messages upstream and therefore
lose the message. Only the device with the conch can send a message and
by implication you only need to be ready to receive downstream messages
if you have sent the conch downstream.