[device] Formalize flash partitions
Create embedded_storage which contains systems that work over any
`embedded_storage::NorFlash` backend i.e. doesn't depend on esp32_*.

In particular I've `FlashPartition` which creates a sort of virtual
flash drive that you can pass to any sub system so it can exclusively
write to it (and cannot write outside of it).

In addition, I've moved the append only style database to
`embedded_storage` along with some of the logic around "A-B" slots.