[app,coord,api] Recover tx signing sessions
On the dart side, unfinished tx signing sessions are displayed on the
top of the wallet tx list. Finished tx signing sessions that are
unbroadcasted are also displayed. We rework the tx details page to allow
signing tx. The send flow redirects to the tx details page after signing
devices are chosen.

On the rust side, we introduce a signal stream for notifying when the
associated signing sessions for a given wallet changes. We introduce
apis for displaying more about signing sessions. We introduce apis for
displaying finished signing sessions and forgetting about them.