[native,app] Remove address metadata
BDK lets us determine which address has been revealed yet or not if we
use it properly. This obviates the need for a separate database table
for now. Instead of revealing spks at the BDK level in order to see them
in the UI we just peek them using the descriptor.