[coord] Introduce UiStack
To replace a Option<Box<dyn UiProtocol>> we have a Vec<Box<dyn
UiProtocol>> so we can have multiple UiProtocols running at once. The
problem this immediately solves is if you need to do a device
consolidation during some other protocol e.g. you are signing and then
you plug a device in recover mode in -- you need to run the
consolidatation which has its own ToUserMessage. This allows you to do both.