[core, coord, app] properly handle share restoration compatibility cases
We had some tracking of `already_held` on both the dart and rust side,
we weren't using it on the rust side. Removed from rust side, but it may
be worth considering if it belongs there instead.

AlreadyGot wasn't referring to the fact that we received this same share,
but rather that we already have received a share from this device.
This commit now handles both.

Also handle the case of a duplicate plugged in device share.