Fix stack overflow in Bip39Framebuf::new
The previous implementation dereferenced a Box<Fb> when creating the
Rc<RefCell<>>, which moved the entire framebuffer onto the stack before
putting it into the Rc. This caused a stack overflow on ESP32C3 with
its limited stack space.

Now we create the Fb directly inside Rc::new(RefCell::new(Fb::new()))
which allows the compiler to optimize and construct it directly on the
heap without ever putting it on the stack.