[core] Allow repeating the same sign requests
We had a system to cache a signature we've just produced such that if we
get asked to sign again we can just return it. But there were checks
that returned errors before the cache was hit. We remove the errors.
This means that attempts to reuse nonces are only caught after the user
has confirmed.