[core] Deprecate UnSave
This mutation was a bit of a nuisance. It was used to remove backup data
on the device after the backup was consolidated. But this was lazy, the
backup data can just be removed after consolidating the key, no need for
this extra mutation.

It's now a noop. If we want to be able to remove backups in the future
we can adda mutation for that, but we have to be very careful to make
sure we actually clear them from the AB slot on the device (and don't
just write this mutation to the log).