[backup] Prefer full fingerprint matches
The proptest hit a pathological case where we would match a fingerprint
with 4 bogus shares where as 3 shares was the correct solution. The 4
bogus shares didn't really match the fingerprint, they were just the
same value so it recovered a threshold 1 poly which auto-matches. Now
secpfun will tell us the degree of the fingerprint match so we can
prefer the best match (and early exit on a total match).